fixed mentioned issues
